[?+] Historical commit. Tried using interface for bits..
but rejected because of huge code overhead. That's glorious C.

Some bits are just switches: want to turn off receiver? Set bit to 0.

So it would be nice if we have Receiver.Switch which is instance of
class TSwitch.

For readability we want many interface functions: IsOn(), IsOff(),
On(), Off(). Maybe even Toggle().

For easy implementation we just need Set() and Get().

  class TSwitch
  {
    public:
      TBool IsOn();
      TBool IsOff();
      void On();
      void Off();
    protected:
      virtual void Get(TBool *) = 0;
      virtual void Set(TBool) = 0;
  };

So IsOn() just calls Set(true). We'll have to define Set() and Get()
for receiver.

It all compiles and runs, but overhead...

For Echo example, code and memory:

Before: 544 9
After: 898 49

And that's the price for ease for developer.

We can't accept that. C is awful.