[^] [Internal] Using Switch class for better code..
.. for more readable outer code.

Prototype of Switch class in previous commit is nice.
But not practical because of C language design.

We've created compromise version of Switch without using
virtual methods:

  bool IsOn()
  void On()
  void Off()

It's three methods we need to define for every switch.
(In prototype it was two. But they were internal and
had less readable interface.)